datasets/grok_submission_2025/model/requirements.txt
onnxruntime>=1.16.0
Pillow>=9.0.0
numpy>=1.21.0
torch>=1.9.0
torchvision>=0.10.0
opencv-python>=4.5.0
albumentations>=1.0.0
scikit-image>=0.18.0
tqdm>=4.60.0